chore(json-api-serializer): first draft with marshmallow-json-api

+++ django_forest/resources/views/detail.py
@@ -1,51 +1,64 @@
 import json
+from datetime import datetime
 
-from django.forms import model_to_dict
+from django.db import models
 from django.http import JsonResponse, HttpResponse
 from django.views import generic
 
-from django_forest.resources.utils import get_model
+from django_forest.resources.utils import get_model, SmartFieldMixin
+from django_forest.utils.json_api_serializer import JsonApiSchema
 
 
-class DetailView(generic.View):
+class DetailView(SmartFieldMixin, generic.View):
+
     def get(self, request, resource, pk, *args, **kwargs):
-        data = {}
+        Model = get_model(resource)
+        if Model is None:
+            return JsonResponse({'message': 'error no model found'}, status=400)
 
-        model = get_model(resource.lower())
-        if model is not None:
-            queryset = model.objects.get(pk=pk)
-            res = model_to_dict(queryset)
-            # TODO
-            data = {
-                'attributes': res,
-                'id': res['id'],
-                'type': resource,
-                'relationships': {}
-            }
+        queryset = Model.objects.get(pk=pk)
 
-        return JsonResponse({'data': data}, safe=False)
+        # handle smart fields
+        self.handle_smart_fields(queryset, resource, Model)
+
+        # json api serializer
+        Schema = JsonApiSchema._registry[f'{resource}Schema']
+        data = Schema().dump(queryset)
+
+        return JsonResponse(data, safe=False)
+
+    def format(self, value, field):
+        # TODO other special fields, put in a mixin
+        if isinstance(field, models.DateTimeField):
+            return datetime.strptime(value, '%Y-%m-%dT%H:%M:%S.%f%z')
+
+        return value
 
     def put(self, request, resource, pk, *args, **kwargs):
         # TODO
+        Model = get_model(resource)
+        if Model is None:
+            return JsonResponse({'message': 'error no model found'}, status=400)
+
         body_unicode = request.body.decode('utf-8')
         body = json.loads(body_unicode)
-        model = get_model(resource)
-        obj = model.objects.get(pk=pk)
+        obj = Model.objects.get(pk=pk)
+        fields = {x.name: x for x in Model._meta.get_fields()}
         for k, v in body['data']['attributes'].items():
-            setattr(obj, k, v)
+            setattr(obj, k, self.format(v, fields[k]))
         obj.save()
-        res = model_to_dict(obj)
-        data = {
-            'attributes': res,
-            'id': res['id'],
-            'type': resource,
-            'relationships': {}
-        }
-        return JsonResponse({'data': data}, safe=False)
+
+        # json api serializer
+        Schema = JsonApiSchema._registry[f'{resource}Schema']
+        data = Schema().dump(obj)
+        return JsonResponse(data, safe=False)
 
     def delete(self, request, resource, pk, *args, **kwargs):
         # TODO
-        model = get_model(resource)
-        obj = model.objects.get(pk=pk)
+        Model = get_model(resource)
+        if Model is None:
+            return JsonResponse({'message': 'error no model found'}, status=400)
+
+        obj = Model.objects.get(pk=pk)
         obj.delete()
         return HttpResponse(status=204)
